feat(05-auth-social): implement google and github social authentications
