+++ src/app/(auth)/sign-up/page.tsx
@@ -1,7 +1,17 @@
-import { SignUpView } from "@/auth/ui/views/sign-up-view";
+import { auth } from "@/lib/auth";
+import { SignUpView } from "@/modules/auth/ui/views/sign-up-view";
+import { headers } from "next/headers";
+import { redirect } from "next/navigation";
 import React from "react";
 
-const Page = () => {
+const Page = async () => {
+  const session = await auth.api.getSession({
+    headers: await headers(),
+  });
+
+  if (!!session) {
+    redirect("/");
+  }
   return <SignUpView />;
 };
