+++ src/app/page.tsx
@@ -1,108 +1,18 @@
-"use client";
+import { auth } from "@/lib/auth";
+import { HomeView } from "@/modules/home/ui/views/home-view";
+import { headers } from "next/headers";
+import { redirect } from "next/navigation";
+import React from "react";
 
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import React, { useState } from "react";
-import { authClient } from "@/lib/auth-client";
-import { Card } from "@/components/ui/card";
+const Page = async () => {
+  const session = await auth.api.getSession({
+    headers: await headers(),
+  });
 
-export default function Home() {
-  const [name, setName] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const { data: session } = authClient.useSession();
-
-  const onSubmit = () => {
-    authClient.signUp.email(
-      {
-        email,
-        name,
-        password,
-      },
-      {
-        onRequest: () => {
-          //show loading
-        },
-        onSuccess: () => {
-          //redirect to the dashboard or sign in page
-          alert("Success");
-        },
-        onError: () => {
-          // display the error message
-          // alert(ctx.error.message);
-          alert("Something went wrong");
-        },
-      }
-    );
-  };
-
-  const onLogin = () => {
-    authClient.signIn.email(
-      {
-        email,
-        password,
-      },
-      {
-        onRequest: () => {
-          //show loading
-        },
-        onSuccess: () => {
-          //redirect to the dashboard or sign in page
-          alert("Success");
-        },
-        onError: () => {
-          // display the error message
-          // alert(ctx.error.message);
-          alert("Something went wrong");
-        },
-      }
-    );
-  };
-
-  if (session) {
-    return (
-      <div className="flex flex-col p-4 gap-y-4">
-        <p className="">Logged in as {session.user.name}</p>
-        <Button onClick={() => authClient.signOut()}>Sign Out</Button>
-      </div>
-    );
+  if (!session) {
+    redirect("/sign-in");
   }
+  return <HomeView />;
+};
 
-  return (
-    <div className="flex flex-col gap-y-8 m-8">
-      <Card className="p-4">
-        <Input
-          placeholder="name"
-          value={name}
-          onChange={(e) => setName(e.target.value)}
-        />
-        <Input
-          placeholder="email"
-          value={email}
-          onChange={(e) => setEmail(e.target.value)}
-        />
-        <Input
-          type="password"
-          placeholder="password"
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-        />
-        <Button onClick={onSubmit}>Create User</Button>
-      </Card>
-      <Card className="p-4">
-        <Input
-          placeholder="email"
-          value={email}
-          onChange={(e) => setEmail(e.target.value)}
-        />
-        <Input
-          type="password"
-          placeholder="password"
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-        />
-        <Button onClick={onLogin}>Log In</Button>
-      </Card>
-    </div>
-  );
-}
+export default Page;
